migrate legacy inference endpoint
